extract tts and recording logic to custom hooks to clean up Chat.tsx
