+++ tts/tts-next-js-chat/src/app/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
-import "./globals.css";
+import "@/globals.css";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
